reworked move code

+++ server/lib/classes/system.inc.php
@@ -936,15 +936,14 @@ class system{
 
         function move($file1, $file2) {
 		if(file_exists($file1) || is_link($file1)) {
-			return copy($file1, $file2);
-	                if(file_exists($file2) || is_link($file2)) {
+			if(copy($file1, $file2)) {
 				return unlink($file1);
 			} else {
 				$app->log("move failed: couldn't move file/link " .$file1." to ".$file2, LOGLEVEL_DEBUG);
 				return false;
 			}
 		} else {
-			$app->log("move failed: " .$file1." doesn't exist.", LOGLEVEL_DEBUG);
+			$app->log("move failed: source " .$file1." doesn't exist.", LOGLEVEL_DEBUG);
 			return false;
 		}
         }
